Use slowdb with random delay with ghost_point test
